errors

+++ src/pages/ContestPage/ContestPage.jsx
@@ -3,6 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { Pencil, Trash2, Eye } from "lucide-react";
 import "./ContestPage.css";
 
+import toast from "react-hot-toast";
+
 import FullScreenLoader from "../../components/FullScreenLoader/FullScreenLoader";
 
 import { getContests, deleteContest } from "../../services/contestService";
@@ -15,32 +17,31 @@ const ContestsPage = () => {
   const [loading, setLoading] = useState(true);
   const [contestToDelete, setContestToDelete] = useState(null);
 
-  async function handleDelete(id) {
-  try {
-    const result = await deleteContest(id);
-    alert(result.message);
+  async function confirmDelete() {
+    if (!contestToDelete) return;
 
-    setContests(prev => prev.filter(contest => contest.id !== id));
+    try {
+      await deleteContest(contestToDelete.id);
 
-  } catch (error) {
-    console.error(error);
-    alert("Erro ao deletar contest");
+      setContests(prev =>
+        prev.filter(c => c.id !== contestToDelete.id)
+      );
+
+      toast.success("Contest deletado com sucesso!");
+      setContestToDelete(null);
+
+    } catch (err) {
+      console.error(err);
+
+      if (err.status === 403) {
+        toast.error("Você não tem permissão para deletar.");
+      } else if (err.status === 404) {
+        toast.error("Contest não encontrado.");
+      } else {
+        toast.error("Erro ao deletar contest.");
+      }
+    }
   }
-}
-
-async function confirmDelete() {
-  try {
-    await deleteContest(contestToDelete.id);
-
-    setContests(prev =>
-      prev.filter(c => c.id !== contestToDelete.id)
-    );
-
-    setContestToDelete(null);
-  } catch (error) {
-    alert("Erro ao deletar contest");
-  }
-}
 
   function formatDateTime(value) {
     if (!value) return "";
@@ -63,7 +64,8 @@ async function confirmDelete() {
       console.log();
 
     } catch (error) {
-      console.error("Erro ao carregar os contests:", error);
+      console.error(error);
+      toast.error("Erro ao carregar contests.");
     } finally {
       setLoading(false);
     }
@@ -149,8 +151,8 @@ async function confirmDelete() {
                           <Pencil size={18} />
                         </button>
 
-                        <button className="btn-icon delete" 
-                          onClick={() => setContestToDelete(contest)} 
+                        <button className="btn-icon delete"
+                          onClick={() => setContestToDelete(contest)}
                           title="Excluir">
                           <Trash2 size={18} />
                         </button>
@@ -199,8 +201,12 @@ async function confirmDelete() {
                   <td>{contest.teamBased ? "Time" : "Individual"}</td>
                   <td>
                     {contest.codeforcesMirrorUrl ?
-                      <a href={contest.codeforcesMirrorUrl}>
-                        Link do contest</a> : <p style={{ color: "#888" }}>Em breve...</p>
+                      <a
+                        href={contest.codeforcesMirrorUrl}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                      >Link do contest</a> 
+                      : <p style={{ color: "#888" }}>Em breve...</p>
                     }
 
                   </td>
@@ -214,8 +220,8 @@ async function confirmDelete() {
                         <Pencil size={18} />
                       </button>
 
-                      <button className="btn-icon delete" 
-                        onClick={() => setContestToDelete(contest)} 
+                      <button className="btn-icon delete"
+                        onClick={() => setContestToDelete(contest)}
                         title="Excluir">
                         <Trash2 size={18} />
                       </button>
@@ -239,32 +245,32 @@ async function confirmDelete() {
 
       </div>
       {contestToDelete && (
-    <div className="modal-overlay">
-      <div className="modal">
-        <h3>Confirmar exclusão</h3>
-        <p>
-          Você tem certeza que deseja deletar{" "}
-          <strong>{contestToDelete.name}</strong>?
-        </p>
-  
-        <div className="modal-actions">
-          <button
-            className="btn-cancel"
-            onClick={() => setContestToDelete(null)}
-          >
-            Cancelar
-          </button>
-  
-          <button
-            className="btn-delete"
-            onClick={confirmDelete}
-          >
-            Deletar
-          </button>
+        <div className="modal-overlay">
+          <div className="modal">
+            <h3>Confirmar exclusão</h3>
+            <p>
+              Você tem certeza que deseja deletar{" "}
+              <strong>{contestToDelete.name}</strong>?
+            </p>
+
+            <div className="modal-actions">
+              <button
+                className="btn-cancel"
+                onClick={() => setContestToDelete(null)}
+              >
+                Cancelar
+              </button>
+
+              <button
+                className="btn-delete"
+                onClick={confirmDelete}
+              >
+                Deletar
+              </button>
+            </div>
+          </div>
         </div>
-      </div>
-    </div>
-  )}
+      )}
     </div>
   );
 };
